Updated into smaller examples
