+++ examples/common_patterns_for_connectors/incremental_sync_strategies/keyset_pagination/configuration.json
@@ -1,0 +1,3 @@
+{
+  "base_url": "http://127.0.0.1:5001/pagination/keyset"
+}
