+++ examples/common_patterns_for_connectors/incremental_sync_strategies/replay_sync/configuration.json
@@ -1,0 +1,4 @@
+{
+  "base_url": "http://127.0.0.1:5001/incremental/replay",
+  "buffer_hours": 2
+}
