+++ examples/common_patterns_for_connectors/incremental_sync_strategies/step_size_sync/configuration.json
@@ -1,0 +1,6 @@
+{
+  "base_url": "http://127.0.0.1:5001/incremental/step",
+  "initial_id": 1,
+  "step_size": 1000,
+  "max_id": 100000
+}
